Virtual results
Introduce a new syntax for storing multiple results from a single
function.

```python
@xun.function()
def f(arg):
    yield even(arg) is arg * 2
    yield odd(arg) is arg * 2 - 1

@f.interface
def even(arg):
    yield from f(arg)

@f.interface
def odd(arg):
    yield from f(arg)
```

The `xun.Function.interface` defines a callable that is reliant on
another function to produce it's result. Which function is specified by
syntax `yield from <call-to-xun-function>`. For example, `yield from f(1)`
indicates that a call to `f(1)` is required to produce a result for the
interface.